+++ src/models/Theme.ts
@@ -1,0 +1,5 @@
+export interface Theme {
+  label: string;
+  uiTheme: string;
+  path: string;
+}