+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the "theming-devtools" extension will be documented in this file.
 
+## [0.0.2] - 2022-02-04
+
+- Fix border colors
+- Allow theme comparison with any other theme you got installed
+
 ## [0.0.1] - 2022-02-04
 
 - Initial release